pyproject.toml
[project]
name = "ai-rival-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.4",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rich>=15.0.0",
]
